pulla stronzo

+++ controller/CAdmin.php
@@ -17,36 +17,39 @@ class CAdmin
      * @throws SmartyException
      */
     static function adminHome(){
+        if(self::isAdminLogged()==true) {
             $pm = FPersistentManager::getInstance();
-            $view=new VAdmin();
-            $image_reported=array();
-            $image_banned=array();
-            $array_banned=$pm->load("Banned",true,FUser::getClass());
-            if(is_object($array_banned)){
-                $array_b=array();
-                $array_b[]=$array_banned;
-            }else{
-                $array_b=$array_banned;
+            $view = new VAdmin();
+            $image_reported = array();
+            $image_banned = array();
+            $array_banned = $pm->load("Banned", true, FUser::getClass());
+            if (is_object($array_banned)) {
+                $array_b = array();
+                $array_b[] = $array_banned;
+            } else {
+                $array_b = $array_banned;
             }
-            $array_reported=$pm->load("Reported",true,FUser::getClass());
-            if(is_object($array_reported)){
-                $array_r=array();
-                $array_r[]=$array_reported;
-            }else{
-                $array_r=$array_reported;
+            $array_reported = $pm->load("Reported", true, FUser::getClass());
+            if (is_object($array_reported)) {
+                $array_r = array();
+                $array_r[] = $array_reported;
+            } else {
+                $array_r = $array_reported;
             }
-            foreach ($array_r as $a){
-                $id=$a->getImageID();
-                $img=$pm->load("IDimage",$id,FImage::getClass());
-                $image_reported[]=$img;
+            foreach ($array_r as $a) {
+                $id = $a->getImageID();
+                $img = $pm->load("IDimage", $id, FImage::getClass());
+                $image_reported[] = $img;
             }
-            if(isset($array_b[0])){
-            foreach ($array_b as $a){
-                $id=$a->getImageID();
-                $img=$pm->load("IDimage",$id,FImage::getClass());
-                $image_banned[]=$img;
-            }}else $array_b=array();
-            $view->adminHomePage($array_b,$array_r,$image_reported,$image_banned);
+            if (isset($array_b[0])) {
+                foreach ($array_b as $a) {
+                    $id = $a->getImageID();
+                    $img = $pm->load("IDimage", $id, FImage::getClass());
+                    $image_banned[] = $img;
+                }
+            } else $array_b = array();
+            $view->adminHomePage($array_b, $array_r, $image_reported, $image_banned);
+        }else header('Location: /logBook/User/home');
     }
 
     static function adminLogout(){
@@ -61,7 +64,7 @@ class CAdmin
      * @throws SmartyException
      */
     static function banUser($userID){
-
+        if(self::isAdminLogged()==true) {
             $pm = FPersistentManager::getInstance();
             $pm->update("Reported",0,$userID,FUser::getClass());
             $resultPost=$pm->load("IDuser",$userID,FPost::getClass());
@@ -88,20 +91,22 @@ class CAdmin
             }
             $pm->update("Banned",1,$userID,FUser::getClass());
             header('Location: /logBook/Admin/adminHome');
+        }else header('Location: /logBook/User/home');
     }
 
     /**
      * @throws SmartyException
      */
     static function ignoreUser($userID){
-
+        if(self::isAdminLogged()==true) {
             $pm = FPersistentManager::getInstance();
             $pm->update("Reported", 0, $userID, FUser::getClass());
             header('Location: /logBook/Admin/adminHome');
+        }else header('Location: /logBook/User/home');
     }
 
     static function restoreUser($userID){
-
+        if(self::isAdminLogged()==true) {
             $pm = FPersistentManager::getInstance();
             $pm->update("Reported", 0, $userID, FUser::getClass());
             $resultPost = $pm->load("IDuser", $userID, FPost::getClass());
@@ -128,13 +133,14 @@ class CAdmin
             }
             $pm->update("Banned", 0, $userID, FUser::getClass());
             header('Location: /logBook/Admin/adminHome');
+        }else header('Location: /logBook/User/home');
     }
 
     /**
      * @throws SmartyException
      */
     static function reportedComments(){
-
+        if(self::isAdminLogged()==true) {
             $pm = FPersistentManager::getInstance();
             $view = new VAdmin;
             $reportedComment = $pm->loadAllDeletedComment();
@@ -148,23 +154,32 @@ class CAdmin
                 $image[] = $img;
             }
             $view->toReportedComments($reportedComment, $author, $image);
-
+        }else header('Location: /logBook/User/home');
     }
 
     static function deleteComment($id){
-        $pm=FPersistentManager::getInstance();
-        $pm->deleteComment($id);
-        $pm->deleteFromCommentReported($id);
-        header('Location: /logBook/Admin/reportedComments');
+        if(self::isAdminLogged()==true) {
+            $pm=FPersistentManager::getInstance();
+            $pm->deleteComment($id);
+            $pm->deleteFromCommentReported($id);
+            header('Location: /logBook/Admin/reportedComments');
+        }else header('Location: /logBook/User/home');
     }
 
     static function ignoreComment($id){
-        $pm=FPersistentManager::getInstance();
-        $pm->deleteFromCommentReported($id);
-        header('Location: /logBook/Admin/reportedComments');
+        if(self::isAdminLogged()==true) {
+            $pm=FPersistentManager::getInstance();
+            $pm->deleteFromCommentReported($id);
+            header('Location: /logBook/Admin/reportedComments');
+        }else header('Location: /logBook/User/home');
     }
 
 
-
+    static function reportedPosts(){
+        if(self::isAdminLogged()==true) {
+            $view = new VAdmin;
+            $view->toReportedPosts();
+        }else header('Location: /logBook/User/home');
+    }
 
 }
